Use custom logger for eradani-connect (#18)     - Move transport init to connection service     - Import same transport for all calls     - Set context "eradani-inc/eradani-connect"
