+++ src/controllers/rpg.ts
@@ -1,19 +1,8 @@
-import configService from 'config';
-const config = configService.get();
-import eradaniConnect from '@eradani-inc/eradani-connect';
-import * as loggerService from '../services/logger';
+import createLogger from 'src/services/logger';
 import TEMPLATE, { TEMPLATEInput, TEMPLATEOutput } from '../models/pgm-template';
+import transport from 'src/services/connection';
 
-const credentials = config.eradaniConnect.credentials;
-const logger = loggerService.createForContext('controllers/rpg');
-
-// Const transport = new eradaniConnect.transports.Odbc(config.eradaniConnect.odbc);
-const transport = new eradaniConnect.transports.Xml(
-    '*LOCAL',
-    credentials.username,
-    credentials.password,
-    config.eradaniConnect.xml
-);
+const logger = createLogger('controllers/rpg');
 
 /**
  * Run the Template program.
